move internal to store/rootmulti
