+++ store/rootmulti/internal/proofs/create.go
@@ -5,7 +5,7 @@ import (
 	"sort"
 
 	ics23 "github.com/confio/ics23/go"
-	sdkmaps "github.com/cosmos/cosmos-sdk/internal/maps"
+	sdkmaps "github.com/cosmos/cosmos-sdk/store/rootmulti/internal/maps"
 )
 
 // TendermintSpec constrains the format from ics23-tendermint (crypto/merkle SimpleProof)
