+++ store/rootmulti/internal/proofs/helpers.go
